requirements.txt
python-decouple==3.8
python-telegram-bot==13.14
redis==5.0.1
